agent-sdk/typescript: rewrite moved-file imports to @synadia-ai/agents
Phase 3 of the TS SDK split. Re-points the imports inside the
host-side files (moved in Phase 2) at the caller package
`@synadia-ai/agents` for everything that stays caller-side: subjects,
errors, validation, NATS-CLI context, byte/version helpers, the
envelope decoder, and the `DecodedAttachment` type. Host-internal
imports (encoder ↔ heartbeat builder ↔ service) keep relative paths.

Caller's `src/index.ts` adjusted in lockstep:
- adds `newInbox` (with `@internal` JSDoc) — the host SDK reuses the
  shared `_INBOX.agents` prefix so caller and host inboxes don't
  diverge.
- adds the decoder-side surface that was previously internal:
  `decodeChunk`, `DecodedChunk`, `DecodedQuery`, `DecodedResponse`,
  `DecodedStatus` (used by the host's `chunk-encoder` round-trip
  tests; matches the plan's "Wire codec — decoder side stays caller").
- swaps `buildHeartbeatPayload` / `encodeHeartbeatPayload` for
  `decodeHeartbeatPayload` to mirror the heartbeat split done in
  Phase 2 (the build/encode pair lives in `@synadia-ai/agent-service`
  now).

The dead `service.ts` bottom-of-file re-exports
(`encodeEnvelope`, `RequestEnvelope`/`RequestAttachment`,
`Chunk`/`ResponseChunk`/`StatusChunk`/`QueryChunk`) are dropped.
They were never reachable from any documented subpath, and the new
package boundary makes `@synadia-ai/agents` (caller) and the host's
`./stream/chunk-encoder.js` re-export the canonical types.

Tree state after this commit: the moved files compile in isolation,
but `client-sdk/typescript/src/index.ts` still has dangling re-exports
of `AgentService` / `encodeChunk` / etc. pointing at files that no
longer exist on the caller side. Phase 4 trims them — the tree only
typechecks end-to-end after Phase 5.
